+++ vz-licencecheck-ui/src/app/app.routes.ts
@@ -3,22 +3,22 @@ import {PolicyGuard} from '@app/_guards/policy.guard';
 import {HomeComponent} from '@app/pages/home/home.component';
 import {verificationResultGuard} from '@app/_guards/verification-result.guard';
 import {scanQrCodeGuard} from '@app/_guards/scan-qr-code.guard';
-import {useCaseGuard} from '@app/_guards/use-case.guard';
 import {mdlFeatureGuard} from '@app/_guards/mdl-feature.guard';
+import {licenseTypeGuard} from '@app/_guards/license-type.guard';
 
 export const routes: Routes = [
   {path: '', component: HomeComponent, pathMatch: 'full'},
   {
-    path: 'use-case',
-    loadComponent: () => import('./pages/use-case/use-case.component').then(m => m.UseCaseComponent),
-    canActivate: [PolicyGuard],
+    path: 'licence-type',
+    loadComponent: () =>
+      import('./pages/licence-type-selector/licence-type-selector.component').then(m => m.LicenceTypeSelectorComponent),
+    canActivate: [PolicyGuard, mdlFeatureGuard],
     pathMatch: 'full'
   },
   {
-    path: 'input-selection',
-    loadComponent: () =>
-      import('./pages/input-selection/input-selection.component').then(m => m.InputSelectionComponent),
-    canActivate: [PolicyGuard, useCaseGuard, mdlFeatureGuard],
+    path: 'use-case',
+    loadComponent: () => import('./pages/use-case/use-case.component').then(m => m.UseCaseComponent),
+    canActivate: [PolicyGuard, licenseTypeGuard],
     pathMatch: 'full'
   },
   {